refactor(test): 픽스처 적용
