+++ src/test/java/kitchenpos/deliveryorder/domain/InMemoryDeliveryOrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.deliveryorder.domain;
 
-import kitchenpos.domain.DeliveryOrderRepository;
 import kitchenpos.order.deliveryorder.domain.DeliveryOrder;
+import kitchenpos.order.deliveryorder.domain.DeliveryOrderRepository;
 
 import java.util.*;
