+++ src/test/java/kitchenpos/eatinorder/application/RestaurantTableServiceTest.java
@@ -1,7 +1,5 @@
 package kitchenpos.eatinorder.application;
 
-import kitchenpos.Fixtures;
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.eatinorder.domain.InMemoryEatInOrderRepository;
 import kitchenpos.eatinorder.domain.InMemoryRestaurantTableRepository;
 import kitchenpos.order.eatinorder.application.RestaurantTableService;
@@ -19,8 +17,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.List;
 import java.util.UUID;
 
-import static kitchenpos.Fixtures.eatInOrder;
-import static kitchenpos.Fixtures.restaurantTable;
+import static kitchenpos.eatinorder.fixture.EatInOrderFixture.eatInOrder;
+import static kitchenpos.eatinorder.fixture.RestaurantTableFixture.restaurantTable;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -44,10 +42,10 @@ class RestaurantTableServiceTest {
         final RestaurantTable actual = restaurantTableService.create(expected);
         assertThat(actual).isNotNull();
         assertAll(
-            () -> assertThat(actual.getId()).isNotNull(),
-            () -> assertThat(actual.getName()).isEqualTo(expected.getName()),
-            () -> assertThat(actual.getNumberOfGuests()).isZero(),
-            () -> assertThat(actual.isOccupied()).isFalse()
+                () -> assertThat(actual.getId()).isNotNull(),
+                () -> assertThat(actual.getName()).isEqualTo(expected.getName()),
+                () -> assertThat(actual.getNumberOfGuests()).isZero(),
+                () -> assertThat(actual.isOccupied()).isFalse()
         );
     }
 
@@ -57,13 +55,13 @@ class RestaurantTableServiceTest {
     void create(final String name) {
         final RestaurantTable expected = createRestaurantTableRequest(name);
         assertThatThrownBy(() -> restaurantTableService.create(expected))
-            .isInstanceOf(IllegalArgumentException.class);
+                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("빈 테이블을 해지할 수 있다.")
     @Test
     void sit() {
-        final UUID orderTableId = restaurantTableRepository.save(Fixtures.restaurantTable(false, 0)).getId();
+        final UUID orderTableId = restaurantTableRepository.save(restaurantTable(false, 0)).getId();
         final RestaurantTable actual = restaurantTableService.sit(orderTableId);
         assertThat(actual.isOccupied()).isTrue();
     }
@@ -71,28 +69,28 @@ class RestaurantTableServiceTest {
     @DisplayName("빈 테이블로 설정할 수 있다.")
     @Test
     void clear() {
-        final UUID orderTableId = restaurantTableRepository.save(Fixtures.restaurantTable(true, 4)).getId();
+        final UUID orderTableId = restaurantTableRepository.save(restaurantTable(true, 4)).getId();
         final RestaurantTable actual = restaurantTableService.clear(orderTableId);
         assertAll(
-            () -> assertThat(actual.getNumberOfGuests()).isZero(),
-            () -> assertThat(actual.isOccupied()).isFalse()
+                () -> assertThat(actual.getNumberOfGuests()).isZero(),
+                () -> assertThat(actual.isOccupied()).isFalse()
         );
     }
 
     @DisplayName("완료되지 않은 주문이 있는 주문 테이블은 빈 테이블로 설정할 수 없다.")
     @Test
     void clearWithUncompletedOrders() {
-        final RestaurantTable restaurantTable = restaurantTableRepository.save(Fixtures.restaurantTable(true, 4));
+        final RestaurantTable restaurantTable = restaurantTableRepository.save(restaurantTable(true, 4));
         final UUID restaurantTableId = restaurantTable.getId();
         eatInOrderRepository.save(eatInOrder(EatInOrderStatus.ACCEPTED, restaurantTable));
         assertThatThrownBy(() -> restaurantTableService.clear(restaurantTableId))
-            .isInstanceOf(IllegalStateException.class);
+                .isInstanceOf(IllegalStateException.class);
     }
 
     @DisplayName("방문한 손님 수를 변경할 수 있다.")
     @Test
     void changeNumberOfGuests() {
-        final UUID orderTableId = restaurantTableRepository.save(Fixtures.restaurantTable(true, 0)).getId();
+        final UUID orderTableId = restaurantTableRepository.save(restaurantTable(true, 0)).getId();
         final RestaurantTable expected = changeNumberOfGuestsRequest(4);
         final RestaurantTable actual = restaurantTableService.changeNumberOfGuests(orderTableId, expected);
         assertThat(actual.getNumberOfGuests()).isEqualTo(4);
@@ -102,19 +100,19 @@ class RestaurantTableServiceTest {
     @ValueSource(ints = -1)
     @ParameterizedTest
     void changeNumberOfGuests(final int numberOfGuests) {
-        final UUID orderTableId = restaurantTableRepository.save(Fixtures.restaurantTable(true, 0)).getId();
+        final UUID orderTableId = restaurantTableRepository.save(restaurantTable(true, 0)).getId();
         final RestaurantTable expected = changeNumberOfGuestsRequest(numberOfGuests);
         assertThatThrownBy(() -> restaurantTableService.changeNumberOfGuests(orderTableId, expected))
-            .isInstanceOf(IllegalArgumentException.class);
+                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("빈 테이블은 방문한 손님 수를 변경할 수 없다.")
     @Test
     void changeNumberOfGuestsInEmptyTable() {
-        final UUID orderTableId = restaurantTableRepository.save(Fixtures.restaurantTable(false, 0)).getId();
+        final UUID orderTableId = restaurantTableRepository.save(restaurantTable(false, 0)).getId();
         final RestaurantTable expected = changeNumberOfGuestsRequest(4);
         assertThatThrownBy(() -> restaurantTableService.changeNumberOfGuests(orderTableId, expected))
-            .isInstanceOf(IllegalStateException.class);
+                .isInstanceOf(IllegalStateException.class);
     }
 
     @DisplayName("주문 테이블의 목록을 조회할 수 있다.")
